requirements.txt
newspaper3k
flask
nltk
bs4
gunicorn